requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
python-dotenv==1.0.0
uuid==1.30
gunicorn==21.2.0 